fix(oauth): corrects signout_path
This change results in redirecting to the login page as it now points to
the correct URI for signing out.

Note: when inspecting network tab in e.g. chromium you will notice CORS
error. I am investigating it but it does not seem to be a blocker for
the OAuth flow to kick in again.

Tested in Chrome and Firefox.