fix: removes cfgmap creation from odh-common manifests
it was conflicting with the same config map created by the plugin.
odh-common, invoked after the plugin, was overwriting service-mesh-refs configmap
with default values which could have been overwritten by the user in the KfDef plugin section.
This led to e.g. workbench not being added to the mesh, as service mesh member was wrongly configured.